refactor: update memory management syscall to use uint64_t for size, add music header, implement spotify program, and enhance command handling

+++ Kernel/idt/syscall_handler.c
@@ -240,7 +240,7 @@ static uint64_t sys_ms_elapsed()
 }
 
 // Memory management syscalls
-static void *sys_malloc(size_t size)
+static void *sys_malloc(uint64_t size)
 {
 	return alloc_memory(get_kernel_memory_manager(), size);
 }
